[Hackability Refactor]  Merge the OpenAI files and rename them to openai_api (#1049)
* Merge the OpenAI files and rename them to openai_api

* Updating import with new file name
